requirements.txt
python-telegram-bot
guesslang